opti: drag image to upload directlly
